+++ src/lib/components/interactables/interact.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-import { TerraDraggableDirective } from './draggable.directive';
-
-@NgModule({
-    imports: [],
-    declarations: [TerraDraggableDirective],
-    exports: [TerraDraggableDirective]
-})
-export class TerraInteractModule {}