+++ CHANGELOG.md
@@ -10,8 +10,11 @@
 ### Breaking Changes
 
 -   **terra-stopwatch** has been removed.
+-   **terra-indicator** has been removed.
 -   **tc-checkbox-group** has been removed.
 -   **terra-slider** has been removed.
+-   **interactables** `TerraInteractModule`, `TerraDraggableDirective` and all related interfaces have been removed. See https://github.com/plentymarkets/terra-components/tree/v11.6.0/src/lib/components/interactables for reference.
+-   **dependencies** removed peer dependency `interactjs` since its no longer used.
 
 # 11.6.0 (12.07.2021)